LPS-119634 Move common styles to FragmentContent
