+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/layout-data-items/FragmentWithControls.js
@@ -12,117 +12,26 @@
  * details.
  */
 
-import classNames from 'classnames';
-import React, {useEffect, useState} from 'react';
+import React from 'react';
 
 import useSetRef from '../../../core/hooks/useSetRef';
 import {
 	LayoutDataPropTypes,
 	getLayoutDataItemPropTypes,
 } from '../../../prop-types/index';
-import {useSelector} from '../../store/index';
-import {getFrontendTokenValue} from '../../utils/getFrontendTokenValue';
-import {getResponsiveConfig} from '../../utils/getResponsiveConfig';
-import loadBackgroundImage from '../../utils/loadBackgroundImage';
 import Topper from '../Topper';
 import FragmentContent from '../fragment-content/FragmentContent';
 
 const FragmentWithControls = React.forwardRef(({item, layoutData}, ref) => {
-	const selectedViewportSize = useSelector(
-		(state) => state.selectedViewportSize
-	);
-
 	const [setRef, itemElement] = useSetRef(ref);
 
-	const itemConfig = getResponsiveConfig(item.config, selectedViewportSize);
-
-	const {
-		backgroundColor,
-		backgroundImage,
-		borderColor,
-		borderRadius,
-		borderWidth,
-		fontFamily,
-		fontSize,
-		fontWeight,
-		height,
-		marginBottom,
-		marginLeft,
-		marginRight,
-		marginTop,
-		maxHeight,
-		maxWidth,
-		minHeight,
-		minWidth,
-		opacity,
-		overflow,
-		paddingBottom,
-		paddingLeft,
-		paddingRight,
-		paddingTop,
-		shadow,
-		textAlign,
-		textColor,
-		width,
-	} = itemConfig.styles;
-
-	const [backgroundImageValue, setBackgroundImageValue] = useState('');
-
-	useEffect(() => {
-		loadBackgroundImage(backgroundImage).then(setBackgroundImageValue);
-	}, [backgroundImage]);
-
-	const style = {};
-
-	style.backgroundColor = getFrontendTokenValue(backgroundColor);
-	style.border = `solid ${borderWidth}px`;
-	style.borderColor = getFrontendTokenValue(borderColor);
-	style.borderRadius = getFrontendTokenValue(borderRadius);
-	style.boxShadow = getFrontendTokenValue(shadow);
-	style.color = getFrontendTokenValue(textColor);
-	style.fontFamily = getFrontendTokenValue(fontFamily);
-	style.fontSize = getFrontendTokenValue(fontSize);
-	style.fontWeight = getFrontendTokenValue(fontWeight);
-	style.height = height;
-	style.maxHeight = maxHeight;
-	style.maxWidth = maxWidth;
-	style.minHeight = minHeight;
-	style.minWidth = minWidth;
-	style.opacity = opacity;
-	style.overflow = overflow;
-	style.width = width;
-
-	if (backgroundImageValue) {
-		style.backgroundImage = `url(${backgroundImageValue})`;
-		style.backgroundPosition = '50% 50%';
-		style.backgroundRepeat = 'no-repeat';
-		style.backgroundSize = 'cover';
-	}
-
 	return (
-		<Topper
-			className={classNames(
-				`mb-${marginBottom}`,
-				`ml-${marginLeft}`,
-				`mr-${marginRight}`,
-				`mt-${marginTop}`,
-				`pb-${paddingBottom}`,
-				`pl-${paddingLeft}`,
-				`pr-${paddingRight}`,
-				`pt-${paddingTop}`,
-				{
-					[textAlign]: textAlign !== 'none',
-				}
-			)}
-			item={item}
-			itemElement={itemElement}
-			layoutData={layoutData}
-			style={style}
-		>
+		<Topper item={item} itemElement={itemElement} layoutData={layoutData}>
 			<FragmentContent
 				elementRef={setRef}
 				fragmentEntryLinkId={item.config.fragmentEntryLinkId}
-				itemId={item.itemId}
+				item={item}
+				withinTopper
 			/>
 		</Topper>
 	);
